requirements.txt
# Optional: Uncomment if you want to use Pydantic for validation
# pydantic>=2.0.0
# requirements.txt
openai
pinecone-client
langchain
langchain-community
pypdf
sentence-transformers
groq